start fixing flush not called

+++ src/main/java/sh/fyz/architect/cache/RedisManager.java
@@ -17,17 +17,23 @@ public class RedisManager {
 
     private final ExecutorService pubSubExecutor = Executors.newCachedThreadPool();
 
+    private boolean isReceiver;
     private boolean isAlive = true;
 
-    private RedisManager(String host, String password, int port, int timeout, int maxConnections) {
+    private RedisManager(String host, String password, int port, int timeout, int maxConnections, boolean receiver) {
         JedisPoolConfig config = new JedisPoolConfig();
         config.setMaxTotal(maxConnections);
         config.setMaxIdle(maxConnections / 2);
         config.setMinIdle(1);
         this.jedisPool = new JedisPool(config, host, port, timeout, password);
         this.objectMapper = new ObjectMapper();
-        this.redisQueueActionPool = new RedisQueueActionPool();
-        jedisPool.getResource().flushAll();
+        this.redisQueueActionPool = new RedisQueueActionPool(receiver);
+        this.isReceiver = receiver;
+        if(receiver) jedisPool.getResource().flushAll();
+    }
+
+    public boolean isReceiver() {
+        return isReceiver;
     }
 
     public JedisPool getJedisPool() {
@@ -43,9 +49,9 @@ public class RedisManager {
         return redisQueueActionPool;
     }
 
-    public static void initialize(String host, String password, int port, int timeout, int maxConnections) {
+    public static void initialize(String host, String password, int port, int timeout, int maxConnections, boolean receiver) {
         if (instance == null) {
-            instance = new RedisManager(host, password, port, timeout, maxConnections);
+            instance = new RedisManager(host, password, port, timeout, maxConnections, receiver);
         } else {
             throw new IllegalStateException("RedisManager is already initialized!");
         }
@@ -81,24 +87,24 @@ public class RedisManager {
     }
 
     public <T> List<T> findAll(String pattern, Class<T> type) {
-        List<T> results = new ArrayList<>();
-        try (Jedis jedis = RedisManager.get().getJedisPool().getResource()) {
-            ScanParams scanParams = new ScanParams().match(pattern).count(100);
-            int cursor = 0;
-            do {
-                ScanResult<String> scanResult = jedis.scan(cursor, scanParams);
-                for (String key : scanResult.getResult()) {
-                    T entity = RedisManager.get().find(key, type);
-                    if (entity != null) {
-                        results.add(entity);
-                    }
+        try (Jedis jedis = jedisPool.getResource()) {
+            List<T> result = new ArrayList<>();
+            for (String key : jedis.keys(pattern)) {
+                String data = jedis.get(key);
+                if (data != null) {
+                    try {
+                        T entity = objectMapper.readValue(data, type);
+                        if (entity != null) {
+                            result.add(entity);
+                        }
+                    } catch (Exception ignored) {}
                 }
-                cursor = scanResult.getCursor();
-            } while (cursor != 0);
+            }
+            return result;
         } catch (Exception e) {
             e.printStackTrace();
+            return null;
         }
-        return results;
     }
 
     public void delete(String key) {
